feat: implement robot picking

+++ src/util/mod.rs
@@ -2,10 +2,10 @@ use std::fmt::{Debug, Formatter};
 use std::ops::Range;
 use colored::Colorize;
 
-#[derive(PartialEq, Clone, Copy)]
+#[derive(PartialEq, Clone, Copy, Eq, Hash)]
 pub struct Coord {
-    pub x: u8,
-    pub y: u8,
+    pub x: usize,
+    pub y: usize,
 }
 
 impl Debug for Coord {
@@ -16,11 +16,11 @@ impl Debug for Coord {
 }
 
 impl Coord {
-    pub fn new(x: u8, y: u8) -> Coord {
+    pub fn new(x: usize, y: usize) -> Coord {
         Coord { x, y }
     }
 
-    pub fn random(range_x: Range<u8>, range_y: Range<u8>) -> Coord {
+    pub fn random(range_x: Range<usize>, range_y: Range<usize>) -> Coord {
         let x = rand::random_range(range_x);
         let y = rand::random_range(range_y);
         Coord { x, y }
